requirements.txt
langchain>=0.3
langgraph>=0.6.1
openai
tiktoken

faiss-cpu

pydantic
pydantic-settings

unstructured
beautifulsoup4
lxml
html2text

python-dotenv
typer

requests
httpx

ipykernel
rich

gradio
langchain-openai
langchain-community

# Testing
pytest
pytest-asyncio
